Dries up research endpoints and tests
