+++ keystone_api/apps/research_products/views.py
@@ -6,7 +6,9 @@ serve as the controller layer in Django's MVC-inspired architecture, bridging
 URLs to business logic.
 """
 
+from django.db import models
 from rest_framework import viewsets
+from rest_framework.exceptions import NotFound
 from rest_framework.permissions import IsAdminUser, IsAuthenticated
 
 from .models import *
@@ -16,51 +18,46 @@ from .serializers import *
 __all__ = ['GrantViewSet', 'PublicationViewSet']
 
 
-class PublicationViewSet(viewsets.ModelViewSet):
-    """Manage metadata for research publications."""
+class BaseAffiliatedViewSet(viewsets.ModelViewSet):
+    """Base viewset for resources filtered by user affiliation."""
 
-    queryset = Publication.objects.all()
-    serializer_class = PublicationSerializer
-    search_fields = ['title', 'abstract', 'journal', 'doi', 'team__name']
+    model = None  # Subclasses must define this
     permission_classes = [IsAuthenticated, IsAdminUser | IsTeamMember]
 
-    def get_queryset(self) -> list[Publication]:
-        """Return a list of publications affiliated with the currently authenticated user."""
+    def get_queryset(self) -> models.QuerySet:
+        """Return a queryset filtered by the user's team affiliation and permissions."""
 
         if self.request.user.is_staff:
-            return self.queryset
+            return self.model.objects.all()
 
-        return Publication.objects.affiliated_with_user(self.request.user)
+        return self.model.objects.affiliated_with_user(self.request.user)
 
-    def get_object(self) -> Grant:
-        """Return the object, applying object-level permissions with explicit 403 on denial."""
+    def get_object(self) -> models.Model:
+        """Return the object and apply object-level permission checks."""
 
-        # Todo: should this wrap in a tryexept that replaces errors with a `raise NotFound("Grant not found.")`?
-        obj = Publication.objects.get(pk=self.kwargs["pk"])
+        try:
+            obj = self.model.objects.get(pk=self.kwargs["pk"])
+
+        except self.model.DoesNotExist:
+            raise NotFound(f"{self.model.__name__} not found.")
+
         self.check_object_permissions(self.request, obj)
         return obj
 
 
-class GrantViewSet(viewsets.ModelViewSet):
+class GrantViewSet(BaseAffiliatedViewSet):
     """Track funding awards and grant information."""
 
+    model = Grant
     queryset = Grant.objects.all()
     serializer_class = GrantSerializer
     search_fields = ['title', 'agency', 'team__name']
-    permission_classes = [IsAuthenticated, IsAdminUser | IsTeamMember]
 
-    def get_queryset(self) -> list[Grant]:
-        """Return a list of grants affiliated with the currently authenticated user."""
 
-        if self.request.user.is_staff:
-            return self.queryset
+class PublicationViewSet(BaseAffiliatedViewSet):
+    """Manage metadata for research publications."""
 
-        return Grant.objects.affiliated_with_user(self.request.user)
-
-    def get_object(self) -> Grant:
-        """Return the object, applying object-level permissions with explicit 403 on denial."""
-
-        # Todo: should this wrap in a tryexept that replaces errors with a `raise NotFound("Grant not found.")`?
-        obj = Grant.objects.get(pk=self.kwargs["pk"])
-        self.check_object_permissions(self.request, obj)
-        return obj
+    model = Publication
+    queryset = Publication.objects.all()
+    serializer_class = PublicationSerializer
+    search_fields = ['title', 'abstract', 'journal', 'doi', 'team__name']
